+++ lib/freeclimb/models/incoming_number_result.rb
@@ -114,7 +114,8 @@ module Freeclimb
         :_alias,
         :region,
         :country,
-        :offnet
+        :offnet,
+        :tfn
       ])
     end