+++ spec/models/call_result_spec.rb
@@ -53,6 +53,20 @@ describe Freeclimb::CallResult do
     end
   end
 
+  describe 'test attribute "date_created_iso"' do
+    it "should work" do
+      instance.date_created_iso = "2022-07-05T15:17:05Z"
+      expect(instance.date_created_iso).to eq("2022-07-05T15:17:05Z")
+    end
+  end
+
+  describe 'test attribute "date_updated_iso"' do
+    it "should work" do
+      instance.date_updated_iso = "2022-07-05T15:17:05Z"
+      expect(instance.date_updated_iso).to eq("2022-07-05T15:17:05Z")
+    end
+  end
+
   describe 'test attribute "call_id"' do
     it "should work" do
       instance.call_id = "TEST_STRING"
@@ -136,6 +150,13 @@ describe Freeclimb::CallResult do
     end
   end
 
+  describe 'test attribute "start_time_iso"' do
+    it "should work" do
+      instance.start_time_iso = "2022-07-05T15:17:05Z"
+      expect(instance.start_time_iso).to eq("2022-07-05T15:17:05Z")
+    end
+  end
+
   describe 'test attribute "connect_time"' do
     it "should work" do
       instance.connect_time = "TEST_STRING"
@@ -143,10 +164,24 @@ describe Freeclimb::CallResult do
     end
   end
 
+  describe 'test attribute "connect_time_iso"' do
+    it "should work" do
+      instance.connect_time_iso = "2022-07-05T15:17:05Z"
+      expect(instance.connect_time_iso).to eq("2022-07-05T15:17:05Z")
+    end
+  end
+
   describe 'test attribute "end_time"' do
     it "should work" do
       instance.end_time = "TEST_STRING"
       expect(instance.end_time).to eq("TEST_STRING")
+    end
+  end
+
+  describe 'test attribute "end_time_iso"' do
+    it "should work" do
+      instance.end_time_iso = "2022-07-05T15:17:05Z"
+      expect(instance.end_time_iso).to eq("2022-07-05T15:17:05Z")
     end
   end
 
@@ -161,6 +196,13 @@ describe Freeclimb::CallResult do
     it "should work" do
       instance.connect_duration = 1
       expect(instance.connect_duration).to eq(1)
+    end
+  end
+
+  describe 'test attribute "audio_stream_duration"' do
+    it "should work" do
+      instance.audio_stream_duration = 1
+      expect(instance.audio_stream_duration).to eq(1)
     end
   end
 
@@ -189,14 +231,24 @@ describe Freeclimb::CallResult do
       expect(instance.answered_by).to eq(Freeclimb::AnsweredBy::MACHINE)
     end
   end
+  describe 'test attribute "caller_name"' do
+    it "should work" do
+      instance.caller_name = "TEST_STRING"
+      expect(instance.caller_name).to eq("TEST_STRING")
+    end
+  end
+
+  describe 'test attribute "web_rtc"' do
+    it "should work" do
+      instance.web_rtc = false
+      expect(instance.web_rtc).to eq(false)
+    end
+  end
+
   describe 'test attribute "subresource_uris"' do
     it "should work" do
-      testObject = Object.new
-      instance.subresource_uris = testObject
-      expect(instance.subresource_uris).to eq(testObject)
-
-      instance.subresource_uris = Object.new
-      expect(instance.subresource_uris).to be_instance_of(Object)
+      instance.subresource_uris = CallResultAllOfSubresourceUris.new
+      expect(instance.subresource_uris).to be_instance_of(CallResultAllOfSubresourceUris)
     end
   end
 
@@ -219,6 +271,10 @@ describe Freeclimb::CallResult do
 
           revision: 1,
 
+          date_created_iso: "2022-07-05T15:17:05Z",
+
+          date_updated_iso: "2022-07-05T15:17:05Z",
+
           call_id: "TS",
 
           parent_call_id: "TS",
@@ -235,19 +291,31 @@ describe Freeclimb::CallResult do
 
           start_time: "TS",
 
+          start_time_iso: "2022-07-05T15:17:05Z",
+
           connect_time: "TS",
 
+          connect_time_iso: "2022-07-05T15:17:05Z",
+
           end_time: "TS",
+
+          end_time_iso: "2022-07-05T15:17:05Z",
 
           duration: 1,
 
           connect_duration: 1,
 
+          audio_stream_duration: 1,
+
           direction: Freeclimb::CallDirection::INBOUND,
 
           answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-          subresource_uris: Object.new,
+          caller_name: "TS",
+
+          web_rtc: true,
+
+          subresource_uris: CallResultAllOfSubresourceUris.new,
 
           application_id: "TS"
         )
@@ -264,6 +332,10 @@ describe Freeclimb::CallResult do
 
           revision: 1,
 
+          date_created_iso: "2022-07-05T15:17:05Z",
+
+          date_updated_iso: "2022-07-05T15:17:05Z",
+
           call_id: "TS",
 
           parent_call_id: "TS",
@@ -280,19 +352,31 @@ describe Freeclimb::CallResult do
 
           start_time: "TS",
 
+          start_time_iso: "2022-07-05T15:17:05Z",
+
           connect_time: "TS",
 
+          connect_time_iso: "2022-07-05T15:17:05Z",
+
           end_time: "TS",
+
+          end_time_iso: "2022-07-05T15:17:05Z",
 
           duration: 1,
 
           connect_duration: 1,
 
+          audio_stream_duration: 1,
+
           direction: Freeclimb::CallDirection::INBOUND,
 
           answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-          subresource_uris: Object.new,
+          caller_name: "TS",
+
+          web_rtc: true,
+
+          subresource_uris: CallResultAllOfSubresourceUris.new,
 
           application_id: "TS",
 
@@ -311,6 +395,10 @@ describe Freeclimb::CallResult do
 
           revision: 1,
 
+          date_created_iso: "2022-07-05T15:17:05Z",
+
+          date_updated_iso: "2022-07-05T15:17:05Z",
+
           call_id: "TS",
 
           parent_call_id: "TS",
@@ -327,19 +415,31 @@ describe Freeclimb::CallResult do
 
           start_time: "TS",
 
+          start_time_iso: "2022-07-05T15:17:05Z",
+
           connect_time: "TS",
 
+          connect_time_iso: "2022-07-05T15:17:05Z",
+
           end_time: "TS",
+
+          end_time_iso: "2022-07-05T15:17:05Z",
 
           duration: 1,
 
           connect_duration: 1,
 
+          audio_stream_duration: 1,
+
           direction: Freeclimb::CallDirection::INBOUND,
 
           answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-          subresource_uris: Object.new,
+          caller_name: "TS",
+
+          web_rtc: true,
+
+          subresource_uris: CallResultAllOfSubresourceUris.new,
 
           application_id: "TS",
 
@@ -360,6 +460,10 @@ describe Freeclimb::CallResult do
 
         revision: 1,
 
+        date_created_iso: "2022-07-05T15:17:05Z",
+
+        date_updated_iso: "2022-07-05T15:17:05Z",
+
         call_id: "TS",
 
         parent_call_id: "TS",
@@ -376,19 +480,31 @@ describe Freeclimb::CallResult do
 
         start_time: "TS",
 
+        start_time_iso: "2022-07-05T15:17:05Z",
+
         connect_time: "TS",
 
+        connect_time_iso: "2022-07-05T15:17:05Z",
+
         end_time: "TS",
+
+        end_time_iso: "2022-07-05T15:17:05Z",
 
         duration: 1,
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-        subresource_uris: Object.new,
+        caller_name: "TS",
+
+        web_rtc: true,
+
+        subresource_uris: CallResultAllOfSubresourceUris.new,
 
         application_id: "TS"
       )
@@ -439,9 +555,15 @@ describe Freeclimb::CallResult do
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
+
+        caller_name: "TS",
+
+        web_rtc: true,
 
         subresource_uris: obj_subresource_uris,
 
@@ -480,9 +602,15 @@ describe Freeclimb::CallResult do
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
+
+        caller_name: "TS",
+
+        web_rtc: true,
 
         subresource_uris: obj_subresource_uris,
 
@@ -525,13 +653,17 @@ describe Freeclimb::CallResult do
 
         connect_duration: 2,
 
+        audio_stream_duration: 2,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-        subresource_uris: Object.new,
+        caller_name: "TS",
 
-        subresource_uris: Object.new,
+        web_rtc: true,
+
+        subresource_uris: CallResultAllOfSubresourceUris.new,
 
         application_id: "TS"
       )
@@ -568,11 +700,17 @@ describe Freeclimb::CallResult do
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: nil,
 
         answered_by: nil,
 
-        subresource_uris: Object.new,
+        caller_name: "ST",
+
+        web_rtc: false,
+
+        subresource_uris: CallResultAllOfSubresourceUris.new,
 
         application_id: "ST"
       )
@@ -591,6 +729,10 @@ describe Freeclimb::CallResult do
 
         revision: 1,
 
+        date_created_iso: "2022-07-05T15:17:05Z",
+
+        date_updated_iso: "2022-07-05T15:17:05Z",
+
         call_id: "TS",
 
         parent_call_id: "TS",
@@ -607,19 +749,31 @@ describe Freeclimb::CallResult do
 
         start_time: "TS",
 
+        start_time_iso: "2022-07-05T15:17:05Z",
+
         connect_time: "TS",
 
+        connect_time_iso: "2022-07-05T15:17:05Z",
+
         end_time: "TS",
+
+        end_time_iso: "2022-07-05T15:17:05Z",
 
         duration: 1,
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-        subresource_uris: Object.new,
+        caller_name: "TS",
+
+        web_rtc: true,
+
+        subresource_uris: CallResultAllOfSubresourceUris.new,
 
         application_id: "TS"
       )
@@ -638,6 +792,10 @@ describe Freeclimb::CallResult do
 
         revision: 1,
 
+        date_created_iso: "2022-07-05T15:17:05Z",
+
+        date_updated_iso: "2022-07-05T15:17:05Z",
+
         call_id: "TS",
 
         parent_call_id: "TS",
@@ -654,19 +812,31 @@ describe Freeclimb::CallResult do
 
         start_time: "TS",
 
+        start_time_iso: "2022-07-05T15:17:05Z",
+
         connect_time: "TS",
 
+        connect_time_iso: "2022-07-05T15:17:05Z",
+
         end_time: "TS",
+
+        end_time_iso: "2022-07-05T15:17:05Z",
 
         duration: 1,
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-        subresource_uris: Object.new,
+        caller_name: "TS",
+
+        web_rtc: true,
+
+        subresource_uris: CallResultAllOfSubresourceUris.new,
 
         application_id: "TS"
       )
@@ -679,6 +849,10 @@ describe Freeclimb::CallResult do
 
         revision: 1,
 
+        date_created_iso: "2022-07-05T15:17:05Z",
+
+        date_updated_iso: "2022-07-05T15:17:05Z",
+
         call_id: "TS",
 
         parent_call_id: "TS",
@@ -695,19 +869,31 @@ describe Freeclimb::CallResult do
 
         start_time: "TS",
 
+        start_time_iso: "2022-07-05T15:17:05Z",
+
         connect_time: "TS",
 
+        connect_time_iso: "2022-07-05T15:17:05Z",
+
         end_time: "TS",
+
+        end_time_iso: "2022-07-05T15:17:05Z",
 
         duration: 1,
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-        subresource_uris: Object.new,
+        caller_name: "TS",
+
+        web_rtc: true,
+
+        subresource_uris: CallResultAllOfSubresourceUris.new,
 
         application_id: "TS"
       )
@@ -726,6 +912,10 @@ describe Freeclimb::CallResult do
 
       revision: 1,
 
+      date_created_iso: "2022-07-05T15:17:05Z",
+
+      date_updated_iso: "2022-07-05T15:17:05Z",
+
       call_id: "TS",
 
       parent_call_id: "TS",
@@ -742,19 +932,31 @@ describe Freeclimb::CallResult do
 
       start_time: "TS",
 
+      start_time_iso: "2022-07-05T15:17:05Z",
+
       connect_time: "TS",
 
+      connect_time_iso: "2022-07-05T15:17:05Z",
+
       end_time: "TS",
+
+      end_time_iso: "2022-07-05T15:17:05Z",
 
       duration: 1,
 
       connect_duration: 1,
 
+      audio_stream_duration: 1,
+
       direction: Freeclimb::CallDirection::INBOUND,
 
       answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-      subresource_uris: Object.new,
+      caller_name: "TS",
+
+      web_rtc: true,
+
+      subresource_uris: CallResultAllOfSubresourceUris.new,
 
       application_id: "TS"
     )
@@ -772,6 +974,14 @@ describe Freeclimb::CallResult do
 
     it "deserializes the data of revision" do
       expect(instance._deserialize("Integer", instance.revision)).to be_a_kind_of(Integer)
+    end
+
+    it "deserializes the data of date_created_iso" do
+      expect(instance._deserialize("Time", instance.date_created_iso)).to be_a_kind_of(Time)
+    end
+
+    it "deserializes the data of date_updated_iso" do
+      expect(instance._deserialize("Time", instance.date_updated_iso)).to be_a_kind_of(Time)
     end
 
     it "deserializes the data of call_id" do
@@ -802,12 +1012,24 @@ describe Freeclimb::CallResult do
       expect(instance._deserialize("String", instance.start_time)).to be_a_kind_of(String)
     end
 
+    it "deserializes the data of start_time_iso" do
+      expect(instance._deserialize("Time", instance.start_time_iso)).to be_a_kind_of(Time)
+    end
+
     it "deserializes the data of connect_time" do
       expect(instance._deserialize("String", instance.connect_time)).to be_a_kind_of(String)
     end
 
+    it "deserializes the data of connect_time_iso" do
+      expect(instance._deserialize("Time", instance.connect_time_iso)).to be_a_kind_of(Time)
+    end
+
     it "deserializes the data of end_time" do
       expect(instance._deserialize("String", instance.end_time)).to be_a_kind_of(String)
+    end
+
+    it "deserializes the data of end_time_iso" do
+      expect(instance._deserialize("Time", instance.end_time_iso)).to be_a_kind_of(Time)
     end
 
     it "deserializes the data of duration" do
@@ -818,8 +1040,20 @@ describe Freeclimb::CallResult do
       expect(instance._deserialize("Integer", instance.connect_duration)).to be_a_kind_of(Integer)
     end
 
+    it "deserializes the data of audio_stream_duration" do
+      expect(instance._deserialize("Integer", instance.audio_stream_duration)).to be_a_kind_of(Integer)
+    end
+
+    it "deserializes the data of caller_name" do
+      expect(instance._deserialize("String", instance.caller_name)).to be_a_kind_of(String)
+    end
+
+    it "deserializes the data of web_rtc" do
+      expect(instance._deserialize("Boolean", instance.web_rtc)).to be_a_kind_of(TrueClass)
+    end
+
     it "deserializes the data of subresource_uris" do
-      expect(instance._deserialize("Object", instance.subresource_uris)).to be_a_kind_of(Object)
+      expect(instance._deserialize("Object", instance.subresource_uris)).to be_a_kind_of(CallResultAllOfSubresourceUris)
     end
 
     it "deserializes the data of application_id" do
@@ -838,6 +1072,10 @@ describe Freeclimb::CallResult do
 
         revision: 1,
 
+        date_created_iso: "2022-07-05T15:17:05Z",
+
+        date_updated_iso: "2022-07-05T15:17:05Z",
+
         call_id: "TS",
 
         parent_call_id: "TS",
@@ -854,19 +1092,31 @@ describe Freeclimb::CallResult do
 
         start_time: "TS",
 
+        start_time_iso: "2022-07-05T15:17:05Z",
+
         connect_time: "TS",
 
+        connect_time_iso: "2022-07-05T15:17:05Z",
+
         end_time: "TS",
+
+        end_time_iso: "2022-07-05T15:17:05Z",
 
         duration: 1,
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-        subresource_uris: Object.new,
+        caller_name: "TS",
+
+        web_rtc: true,
+
+        subresource_uris: CallResultAllOfSubresourceUris.new,
 
         application_id: "TS"
       )
@@ -885,6 +1135,10 @@ describe Freeclimb::CallResult do
 
         revision: 1,
 
+        date_created_iso: "2022-07-05T15:17:05Z",
+
+        date_updated_iso: "2022-07-05T15:17:05Z",
+
         call_id: "TS",
 
         parent_call_id: "TS",
@@ -901,26 +1155,38 @@ describe Freeclimb::CallResult do
 
         start_time: "TS",
 
+        start_time_iso: "2022-07-05T15:17:05Z",
+
         connect_time: "TS",
 
+        connect_time_iso: "2022-07-05T15:17:05Z",
+
         end_time: "TS",
+
+        end_time_iso: "2022-07-05T15:17:05Z",
 
         duration: 1,
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
 
-        subresource_uris: Object.new,
+        caller_name: "TS",
+
+        web_rtc: true,
+
+        subresource_uris: CallResultAllOfSubresourceUris.new,
 
         application_id: "TS"
       )
       expect(instance.to_hash).to be_a_kind_of(Hash)
     end
     it "creates equal hash for two equal objects" do
-      obj = Object.new
+      obj = CallResultAllOfSubresourceUris.new
 
       instance_1 = Freeclimb::CallResult.new(
         uri: "TS",
@@ -931,6 +1197,10 @@ describe Freeclimb::CallResult do
 
         revision: 1,
 
+        date_created_iso: "2022-07-05T15:17:05Z",
+
+        date_updated_iso: "2022-07-05T15:17:05Z",
+
         call_id: "TS",
 
         parent_call_id: "TS",
@@ -947,17 +1217,29 @@ describe Freeclimb::CallResult do
 
         start_time: "TS",
 
+        start_time_iso: "2022-07-05T15:17:05Z",
+
         connect_time: "TS",
 
+        connect_time_iso: "2022-07-05T15:17:05Z",
+
         end_time: "TS",
+
+        end_time_iso: "2022-07-05T15:17:05Z",
 
         duration: 1,
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
+
+        caller_name: "TS",
+
+        web_rtc: true,
 
         subresource_uris: obj,
 
@@ -972,6 +1254,10 @@ describe Freeclimb::CallResult do
 
         revision: 1,
 
+        date_created_iso: "2022-07-05T15:17:05Z",
+
+        date_updated_iso: "2022-07-05T15:17:05Z",
+
         call_id: "TS",
 
         parent_call_id: "TS",
@@ -988,17 +1274,29 @@ describe Freeclimb::CallResult do
 
         start_time: "TS",
 
+        start_time_iso: "2022-07-05T15:17:05Z",
+
         connect_time: "TS",
 
+        connect_time_iso: "2022-07-05T15:17:05Z",
+
         end_time: "TS",
+
+        end_time_iso: "2022-07-05T15:17:05Z",
 
         duration: 1,
 
         connect_duration: 1,
 
+        audio_stream_duration: 1,
+
         direction: Freeclimb::CallDirection::INBOUND,
 
         answered_by: Freeclimb::AnsweredBy::HUMAN,
+
+        caller_name: "TS",
+
+        web_rtc: true,
 
         subresource_uris: obj,
 
@@ -1018,6 +1316,10 @@ describe Freeclimb::CallResult do
 
       revision: 1,
 
+      date_created_iso: "2022-07-05T15:17:05Z",
+
+      date_updated_iso: "2022-07-05T15:17:05Z",
+
       call_id: "TS",
 
       parent_call_id: "TS",
@@ -1034,17 +1336,29 @@ describe Freeclimb::CallResult do
 
       start_time: "TS",
 
+      start_time_iso: "2022-07-05T15:17:05Z",
+
       connect_time: "TS",
 
+      connect_time_iso: "2022-07-05T15:17:05Z",
+
       end_time: "TS",
+
+      end_time_iso: "2022-07-05T15:17:05Z",
 
       duration: 1,
 
       connect_duration: 1,
 
+      audio_stream_duration: 1,
+
       direction: Freeclimb::CallDirection::INBOUND,
 
       answered_by: Freeclimb::AnsweredBy::HUMAN,
+
+      caller_name: "TS",
+
+      web_rtc: true,
 
       application_id: "TS"
     )
@@ -1059,6 +1373,12 @@ describe Freeclimb::CallResult do
     end
     it "returns revision in the form of hash" do
       expect(instance._to_hash(instance.revision)).to eq(instance.revision)
+    end
+    it "returns date_created_iso in the form of hash" do
+      expect(instance._to_hash(instance.date_created_iso)).to eq(instance.date_created_iso)
+    end
+    it "returns date_updated_iso in the form of hash" do
+      expect(instance._to_hash(instance.date_updated_iso)).to eq(instance.date_updated_iso)
     end
     it "returns call_id in the form of hash" do
       expect(instance._to_hash(instance.call_id)).to eq(instance.call_id)
@@ -1084,11 +1404,20 @@ describe Freeclimb::CallResult do
     it "returns start_time in the form of hash" do
       expect(instance._to_hash(instance.start_time)).to eq(instance.start_time)
     end
+    it "returns start_time_iso in the form of hash" do
+      expect(instance._to_hash(instance.start_time_iso)).to eq(instance.start_time_iso)
+    end
     it "returns connect_time in the form of hash" do
       expect(instance._to_hash(instance.connect_time)).to eq(instance.connect_time)
     end
+    it "returns connect_time_iso in the form of hash" do
+      expect(instance._to_hash(instance.connect_time_iso)).to eq(instance.connect_time_iso)
+    end
     it "returns end_time in the form of hash" do
       expect(instance._to_hash(instance.end_time)).to eq(instance.end_time)
+    end
+    it "returns end_time_iso in the form of hash" do
+      expect(instance._to_hash(instance.end_time_iso)).to eq(instance.end_time_iso)
     end
     it "returns duration in the form of hash" do
       expect(instance._to_hash(instance.duration)).to eq(instance.duration)
@@ -1096,11 +1425,20 @@ describe Freeclimb::CallResult do
     it "returns connect_duration in the form of hash" do
       expect(instance._to_hash(instance.connect_duration)).to eq(instance.connect_duration)
     end
+    it "returns audio_stream_duration in the form of hash" do
+      expect(instance._to_hash(instance.audio_stream_duration)).to eq(instance.audio_stream_duration)
+    end
     it "returns direction in the form of hash" do
       expect(instance._to_hash(instance.direction)).to eq(instance.direction)
     end
     it "returns answered_by in the form of hash" do
       expect(instance._to_hash(instance.answered_by)).to eq(instance.answered_by)
+    end
+    it "returns caller_name in the form of hash" do
+      expect(instance._to_hash(instance.caller_name)).to eq(instance.caller_name)
+    end
+    it "returns web_rtc in the form of hash" do
+      expect(instance._to_hash(instance.web_rtc)).to eq(instance.web_rtc)
     end
     it "returns subresource_uris in the form of hash" do
       expect(instance._to_hash(instance.subresource_uris)).to eq(instance.subresource_uris)